requirements.txt
streamlit==1.25.0
openai-whisper
ffmpeg-python
torch>=1.9.0
openai==0.27.8